# backend/main-service/openapi.yaml
openapi: 3.0.3
info:
  title: Educational Programs API
  description: API for managing educational institutions and their study programs with full CRUD operations, search, and filtering capabilities
  version: 1.0.0
...
paths:
...
  /degree-programs:
    get:
      tags:
        - Degree Programs
      summary: List all degree programs
      operationId: listDegreePrograms
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: title
          in: query
          schema:
            type: string
        - name: stream
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            default: title:asc
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DegreeProgram"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Degree Programs
      summary: Create a new degree program
      operationId: createDegreeProgram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DegreeProgramCreate"
      responses:
        "201":
          description: Degree program created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegreeProgram"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /degree-programs/{degreeProgramId}:
    get:
      tags:
        - Degree Programs
      summary: Get degree program by ID
      operationId: getDegreeProgram
      parameters:
        - name: degreeProgramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegreeProgram"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - Degree Programs
      summary: Update degree program
      operationId: updateDegreeProgram
      parameters:
        - name: degreeProgramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DegreeProgramUpdate"
      responses:
        "200":
          description: Degree program updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegreeProgram"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags:
        - Degree Programs
      summary: Partially update degree program
      operationId: patchDegreeProgram
      parameters:
        - name: degreeProgramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DegreeProgramPatch"
      responses:
        "200":
          description: Degree program updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegreeProgram"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Degree Programs
      summary: Delete degree program
      operationId: deleteDegreeProgram
      parameters:
        - name: degreeProgramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Degree program deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /course-offerings:
    get:
      tags:
        - Course Offerings
      summary: List all course offerings
      operationId: listCourseOfferings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: degree_program_id
          in: query
          schema:
            type: string
        - name: university_id
          in: query
          schema:
            type: string
        - name: academic_year
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            default: _id:asc
        - name: stream
          in: query
          schema:
            type: string
          description: Filter by A/L stream
        - name: degree
          in: query
          schema:
            type: string
          description: Filter by degree/program title
        - name: university
          in: query
          schema:
            type: string
          description: Filter by university name
        - name: aptitude
          in: query
          schema:
            type: boolean
          description: Filter by aptitude test requirement
        - name: medium
          in: query
          schema:
            type: string
          description: Filter by medium of instruction
        - name: district
          in: query
          schema:
            type: string
          description: Filter by applicant's district
        - name: z_score
          in: query
          schema:
            type: number
          description: Filter by applicant's Z-score
        - name: eligible_only
          in: query
          schema:
            type: boolean
          description: Return only courses the applicant is eligible for
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourseOffering"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Course Offerings
      summary: Create a new course offering
      operationId: createCourseOffering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseOfferingCreate"
      responses:
        "201":
          description: Course offering created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOffering"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /course-offerings/{courseOfferingId}:
    get:
      tags:
        - Course Offerings
      summary: Get course offering by ID
      operationId: getCourseOffering
      parameters:
        - name: courseOfferingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOffering"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - Course Offerings
      summary: Update course offering
      operationId: updateCourseOffering
      parameters:
        - name: courseOfferingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseOfferingUpdate"
      responses:
        "200":
          description: Course offering updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOffering"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags:
        - Course Offerings
      summary: Partially update course offering
      operationId: patchCourseOffering
      parameters:
        - name: courseOfferingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseOfferingPatch"
      responses:
        "200":
          description: Course offering updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOffering"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Course Offerings
      summary: Delete course offering
      operationId: deleteCourseOffering
      parameters:
        - name: courseOfferingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Course offering deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    DegreeProgram:
      type: object
      required:
        - _id
        - title
        - stream
        - duration_years
        - medium_of_instruction
        - al_requirements
        - ol_requirements
        - aptitude_test_required
      properties:
        _id:
          type: string
        title:
          type: string
        stream:
          type: string
        duration_years:
          type: integer
        medium_of_instruction:
          type: array
          items:
            type: string
        al_requirements:
          type: object
          properties:
            logic_type:
              type: string
            minimum_grade_required:
              type: string
            subjects:
              type: array
              items:
                type: string
            special_notes:
              type: string
        ol_requirements:
          type: array
          items:
            type: object
            properties:
              subject:
                type: string
              min_grade:
                type: string
        aptitude_test_required:
          type: boolean
    DegreeProgramCreate:
      allOf:
        - $ref: "#/components/schemas/DegreeProgram"
    DegreeProgramUpdate:
      allOf:
        - $ref: "#/components/schemas/DegreeProgram"
    DegreeProgramPatch:
      type: object
      properties:
        title:
          type: string
        stream:
          type: string
        duration_years:
          type: integer
        medium_of_instruction:
          type: array
          items:
            type: string
        al_requirements:
          type: object
          properties:
            logic_type:
              type: string
            minimum_grade_required:
              type: string
            subjects:
              type: array
              items:
                type: string
            special_notes:
              type: string
        ol_requirements:
          type: array
          items:
            type: object
            properties:
              subject:
                type: string
              min_grade:
                type: string
        aptitude_test_required:
          type: boolean
    CourseOffering:
      type: object
      required:
        - _id
        - degree_program_id
        - university_id
        - academic_year
        - proposed_intake
        - cutoff_marks
      properties:
        _id:
          type: string
        degree_program_id:
          type: string
        university_id:
          type: string
        academic_year:
          type: string
        proposed_intake:
          type: integer
        cutoff_marks:
          type: object
          additionalProperties:
            type: number
    CourseOfferingCreate:
      allOf:
        - $ref: "#/components/schemas/CourseOffering"
    CourseOfferingUpdate:
      allOf:
        - $ref: "#/components/schemas/CourseOffering"
    CourseOfferingPatch:
      type: object
      properties:
        degree_program_id:
          type: string
        university_id:
          type: string
        academic_year:
          type: string
        proposed_intake:
          type: integer
        cutoff_marks:
          type: object
          additionalProperties:
            type: number
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []